general: Initial commit
This is still a work in progress! No interaction yet and unwraps
everywhere! Proper error handling will probably come next.
The playfield decorative sprites are laid out in a haphazard way
for now. This is very prototypical and all subject to change!

There are no image files in the resources in this commit; I'm
using placeholder assets from the source game for now which are
obviously illegal to distribute, so I haven't included them.